requirements.txt
asgiref==3.8.1
attrs==24.2.0
black==24.8.0
cattrs==23.2.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
Django==5.1
django-database-url==1.0.3
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-login-required-middleware==0.9.0
django-phonenumber-field==8.0.0
django-schema-graph==3.1.0
Faker==28.0.0
gcloud==0.18.3
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.8
jwcrypto==1.5.6
mypy-extensions==1.0.0
oauth2client==4.1.3
packaging==24.1
pathspec==0.12.1
phonenumbers==8.13.44
pillow==10.4.0
platformdirs==4.2.2
protobuf==5.28.1
psycopg==3.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.20.0
pyparsing==3.1.4
Pyrebase4==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jwt==4.1.0
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==1.26.20